Added help, TODO. Improved verbosity, parallelism.
Also added a number of "road-map" style comments for
improvements. Cut out "run_on_finish" parallel lib code
and dropped it into its own function. Added a "debug"
argument which currently does nothing. Added
notifications for the closing of a filehandle when
verbosity is requested from the user. Tried parallelism
with the line-by-line log parsing, but the main data
structure was so slammed with write requests that
performance came to a halt. It seems that child threads
report information to their parents by writing data
structure references to disk, and requiring disk I/O
for each one of the millions of log lines was way too
much for the drive to handle. Searching for a better way.